Allow to put users in system groups (closes #2)
Add a default user and group to main vars file. They act as
"minimum ID" marks so we can identify custom users and groups.
The hard-coded IDs in the tasks have been replaced.

Add a cleaning script to remove custom users from system
groups. This is the missing piece that enables us to allow
using system groups in the main vars file. There is no
mechanism in ansible for this task.
The task running this cleaning script needs to run first,
i.e. before cleaning passwd. It needs the passwd as reference
for the users that have to be removed from system groups.